+++ agglayer/client_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/0xPolygon/cdk-rpc/rpc"
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 )
 
@@ -24,6 +25,30 @@ func TestExploratoryClient(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, lastCert)
 	fmt.Printf("LastCert: %s", lastCert.String())
+}
+
+func TestExploratoryGetCertificateHeader(t *testing.T) {
+	t.Skip("This test is exploratory and should be skipped")
+	aggLayerClient := NewAggLayerClient("http://localhost:32796")
+	certificateID := common.HexToHash("0xf153e75e24591432ac5deafaeaafba3fec0fd851261c86051b9c0d540b38c369")
+	certificateHeader, err := aggLayerClient.GetCertificateHeader(certificateID)
+	require.NoError(t, err)
+	fmt.Print(certificateHeader)
+}
+func TestExploratoryGetEpochConfiguration(t *testing.T) {
+	t.Skip("This test is exploratory and should be skipped")
+	aggLayerClient := NewAggLayerClient("http://localhost:32796")
+	clockConfig, err := aggLayerClient.GetEpochConfiguration()
+	require.NoError(t, err)
+	fmt.Print(clockConfig)
+}
+
+func TestExploratoryGetLatestKnownCertificateHeader(t *testing.T) {
+	t.Skip("This test is exploratory and should be skipped")
+	aggLayerClient := NewAggLayerClient("http://localhost:32781")
+	cert, err := aggLayerClient.GetLatestKnownCertificateHeader(1)
+	require.NoError(t, err)
+	fmt.Print(cert)
 }
 
 func TestGetEpochConfigurationResponseWithError(t *testing.T) {
@@ -80,6 +105,18 @@ func TestGetEpochConfigurationOkResponse(t *testing.T) {
 	}, *clockConfig)
 }
 
+func TestGetLatestKnownCertificateHeaderOkResponse(t *testing.T) {
+	sut := NewAggLayerClient(testURL)
+	response := rpc.Response{
+		Result: []byte(`{"network_id":1,"height":0,"epoch_number":223,"certificate_index":0,"certificate_id":"0xf9179d2fbe535814b5a14496e2eed474f49c6131227a9dfc5d2d8caf9e212054","new_local_exit_root":"0x7ae06f4a5d0b6da7dd4973fb6ef40d82c9f2680899b3baaf9e564413b59cc160","metadata":"0x00000000000000000000000000000000000000000000000000000000000001a7","status":"Settled"}`),
+	}
+	jSONRPCCall = func(url, method string, params ...interface{}) (rpc.Response, error) {
+		return response, nil
+	}
+	cert, err := sut.GetLatestKnownCertificateHeader(1)
+	require.NotNil(t, cert)
+	require.NoError(t, err)
+}
 func TestGetLatestKnownCertificateHeaderErrorResponse(t *testing.T) {
 	sut := NewAggLayerClient(testURL)
 	jSONRPCCall = func(url, method string, params ...interface{}) (rpc.Response, error) {
